Рефакторит модуль utils
