+++ src/utils/film.js
@@ -1,0 +1,3 @@
+export const formatRating = (rating) => rating.toFixed(1);
+
+export const formatItems = (items) => items.join(', ');
