+++ src/views/film-details.js
@@ -1,6 +1,7 @@
 import AbstractView from './abstract.js';
 import { ClassName } from '../const.js';
-import { getReleaseDate, getRuntime, formatRating, formatItems } from '../utils.js';
+import { formatRating, formatItems } from '../utils/film.js';
+import { getReleaseDate, getRuntime } from '../utils/date.js';
 
 const setActiveClassName = (condition) => condition ? ClassName.FILM_DETAILS_CONTROL_ACTIVE : '';
